+++ CHANGELOG.md
@@ -56,8 +56,12 @@
 * Go: Add Function Stats ([#3526](https://github.com/valkey-io/valkey-glide/pull/3526))
 * Go: Add Function Delete ([#3603](https://github.com/valkey-io/valkey-glide/pull/3603))
 * Go: Add Function Kill ([#3604](https://github.com/valkey-io/valkey-glide/pull/3604))
+* Go: Script eval/load([#3605](https://github.com/valkey-io/valkey-glide/pull/3605))
 * Benchmarks: Fix rust benchmark latencies calculation
+* Go: Add Script Exists ([#3649](https://github.com/valkey-io/valkey-glide/pull/3649))
 * Python: Moved the tests folder to be under the root folder ([#3661](https://github.com/valkey-io/valkey-glide/pull/3661))
+* Go: Add script flush command([#3662](https://github.com/valkey-io/valkey-glide/pull/3662))
+* Go: Add Script Kill ([#3670](https://github.com/valkey-io/valkey-glide/pull/3670))
 
 #### Breaking Changes